Ftr: add new-arc related packages
