+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
   "common/logger",
   "common/utils",
-  "xds",
+  "common/extention",
   "registry/zookeeper",
   "registry/nacos",
   "metadata",
@@ -13,7 +13,10 @@ members = [
   "dubbo-build",
   "remoting/net",
   "remoting/http",
+  "remoting/h2",
   "remoting/zookeeper",
+  "remoting/exchange",
+  "remoting/xds",
   "protocol/dubbo2",
   "protocol/protocol",
   "protocol/triple"
@@ -36,10 +39,12 @@ serde_json = "1"
 urlencoding = "2.1.2"
 logger = {path="./common/logger"}
 utils = {path="./common/utils"}
-net = {path="./remoting/net"}
+remoting-net = {path="./remoting/net"}
 protocol = {path="./protocol/protocol"}
-dubbo2 = {path="./protocol/dubbo2"}
-triple = {path="./protocol/triple"}
+protocol-dubbo2 = {path="./protocol/dubbo2"}
+protocol-triple = {path="./protocol/triple"}
+registry-zookeeper = {path="./registry/zookeeper"}
+registry-nacos = {path="./registry/nacos"}
 anyhow = "1.0.66"
 dubbo = { path = "./dubbo/" }
 bb8 = "0.8.0" # A connecton pool based on tokio
