+++ examples/echo/Cargo.toml
@@ -28,13 +28,13 @@ prost-derive = {version = "0.10", optional = true}
 prost = "0.10.4"
 async-trait = "0.1.56"
 tokio-stream = "0.1"
-logger = {path="../../common/logger"}
+logger.workspace=true
 
 hyper = { version = "0.14.19", features = ["full"]}
 
 dubbo = {path = "../../dubbo", version = "0.3.0" }
 dubbo-config = {path = "../../config", version = "0.3.0" }
-dubbo-registry-zookeeper = {path = "../../registry/zookeeper", version = "0.3.0" }
+registry-zookeeper.workspace=true
 
 [build-dependencies]
 dubbo-build = {path = "../../dubbo-build", version = "0.3.0" }
