+++ examples/greeter/Cargo.toml
@@ -22,17 +22,17 @@ path = "src/greeter/client.rs"
 [dependencies]
 http = "0.2"
 http-body = "0.4.4"
-futures-util = {version = "0.3", default-features = false}
-tokio = { version = "1.0", features = [ "rt-multi-thread", "time", "fs", "macros", "net", "signal"] }
-prost-derive = {version = "0.10", optional = true}
+futures-util = { version = "0.3", default-features = false }
+tokio = { version = "1.0", features = ["rt-multi-thread", "time", "fs", "macros", "net", "signal"] }
+prost-derive = { version = "0.10", optional = true }
 prost = "0.10.4"
 async-trait = "0.1.56"
 tokio-stream = "0.1"
-logger = {path="../../common/logger"}
-dubbo = {path = "../../dubbo", version = "0.3.0" }
-dubbo-config = {path = "../../config", version = "0.3.0" }
-dubbo-registry-zookeeper = {path = "../../registry/zookeeper", version = "0.3.0" }
-dubbo-registry-nacos = {path = "../../registry/nacos", version = "0.3.0" }
+logger = { path = "../../common/logger" }
+dubbo = { path = "../../dubbo", version = "0.3.0" }
+dubbo-config = { path = "../../config", version = "0.3.0" }
+registry-zookeeper.workspace = true
+registry-nacos.workspace = true
 
 [build-dependencies]
-dubbo-build = {path = "../../dubbo-build", version = "0.3.0" }
+dubbo-build = { path = "../../dubbo-build", version = "0.3.0" }
