+++ examples/greeter/src/greeter/server.rs
@@ -24,7 +24,6 @@ use tokio_stream::wrappers::ReceiverStream;
 
 use dubbo::{codegen::*, Dubbo};
 use dubbo_config::RootConfig;
-use dubbo_registry_zookeeper::zookeeper_registry::ZookeeperRegistry;
 use logger::{
     tracing::{info, span},
     Level,
@@ -33,6 +32,7 @@ use protos::{
     greeter_server::{register_server, Greeter},
     GreeterReply, GreeterRequest,
 };
+use registry_zookeeper::ZookeeperRegistry;
 
 pub mod protos {
     #![allow(non_camel_case_types)]
