+++ protocol/README.md
@@ -1,0 +1,4 @@
+/protocol
+    /protocol # define protocol abstract layer
+    /dubbo2   # for dubbo2 protocol, hessian2 codec as default
+    /triple   # for triple protocol
