+++ protocol/triple/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "triple"
+name = "protocol-triple"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-net.workspace = true
+remoting-net.workspace = true
 protocol.workspace = true
