+++ registry/README.md
@@ -1,0 +1,3 @@
+/registry
+    /nacos      # nacos registry and servicediscovery implementation
+    /zookeeper  # zookeeper registry and servicediscovery implementation
