+++ registry/nacos/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dubbo-registry-nacos"
+name = "registry-nacos"
 version = "0.3.0"
 edition = "2021"
 license = "Apache-2.0"
