+++ registry/zookeeper/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dubbo-registry-zookeeper"
+name = "registry-zookeeper"
 version = "0.3.0"
 edition = "2021"
 license = "Apache-2.0"
