+++ remoting/README.md
@@ -1,0 +1,6 @@
+/remoting
+    /net   # for dubbo2
+    /http  # on top of net, for hessian/rest/grpc/triple protocol
+    /exchange # convert tcp/http to invocation or invoker for protocol layer 
+    /zookeeper # zk client for registry and config center
+    /nacos  # nacos client for registry and config center
