+++ remoting/exchange/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "protocol-dubbo2"
+name = "remoting-exchange"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-remoting-net.workspace = true
-protocol.workspace = true
