+++ remoting/h2/Cargo.toml
@@ -1,10 +1,7 @@
 [package]
-name = "xds"
-version = "0.3.0"
+name = "remoting-h2"
+version = "0.1.0"
 edition = "2021"
-license = "Apache-2.0"
-description = "dubbo-rust-xds"
-repository = "https://github.com/apache/dubbo-rust.git"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
