+++ remoting/http/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "http"
+name = "remoting-http"
 version = "0.1.0"
 edition = "2021"
