+++ remoting/xds/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
-name = "protocol-dubbo2"
-version = "0.1.0"
+name = "xds"
+version = "0.3.0"
 edition = "2021"
+license = "Apache-2.0"
+description = "remoting-xds"
+repository = "https://github.com/apache/dubbo-rust.git"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-remoting-net.workspace = true
-protocol.workspace = true
